+++ http-client/Network/HTTP/Client/Response.hs
@@ -11,6 +11,8 @@ module Network.HTTP.Client.Response
 import Data.ByteString (ByteString)
 import qualified Data.ByteString.Char8 as S8
 import qualified Data.ByteString.Lazy as L
+import Data.CaseInsensitive (mk)
+import Control.Arrow (second)
 
 import Data.Monoid (mempty)
 
@@ -87,7 +89,7 @@ getResponse :: ConnRelease
 getResponse connRelease timeout' req@(Request {..}) conn cont = do
     StatusHeaders s version hs <- parseStatusHeaders conn timeout' cont
     let mcl = lookup "content-length" hs >>= readDec . S8.unpack
-        isChunked = ("transfer-encoding", "chunked") `elem` hs
+        isChunked = ("transfer-encoding", mk "chunked") `elem` map (second mk) hs
 
         -- should we put this connection back into the connection manager?
         toPut = Just "close" /= lookup "connection" hs && version > W.HttpVersion 1 0